+++ sopalin/coeftab_zdiff.c
@@ -162,7 +162,7 @@ coeftab_zdiff( const SolverMatrix *solvA, SolverMatrix *solvB )
     for(cblknum=0; cblknum<solvA->cblknbr; cblknum++, cblkA++, cblkB++) {
         rc += coeftab_zdiffcblk( cblkA, cblkB );
         if ( rc != saved_rc ){
-            fprintf(stderr, "CBLK %ld was not correctly compressed\n", cblknum);
+            fprintf(stderr, "CBLK %ld was not correctly compressed\n", (long)cblknum);
             saved_rc = rc;
         }
     }
